requirements.txt
matplotlib==3.9.2
numpy==2.1.2
openai==1.51.1
python-dotenv==1.0.1
anthropic==0.37.1
google-generativeai==0.8.3